Remarks.....
Hello, I had trouble with calculating the Mean and SD for every row, in step 2.
For the Mean, I got the Mean of row 1 applied to all the other rows, so 1 value.
Using rowwise() solved this problem.
For the SD, I got only NA's. I tried a lot.The only thing that worked was a c(V1,V2,V3,,,,,,,,) vector.
To type 128 V-values was a little bit to much. So I did only 10 V-columns, as you can see above.
But it worked well. That was all. Thank you for reading. Bye.